refactor post integration
